apps/runtime/pyproject.toml
[project]
name = "wireml-runtime"
version = "0.1.0"
description = "Local Python runtime for WireML — auto-detects CUDA / MPS / MLX / ROCm / DirectML."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "WireML Contributors" }]

dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.4",
    "numpy>=1.26,<2.3",
    "pillow>=11.0.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.46.0",
    "accelerate>=1.2.0",
    "onnxruntime>=1.20.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
mlx = ["mlx>=0.20.0"]
rocm = []
directml = ["onnxruntime-directml>=1.20.0"]
cuda = []
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.1",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
]

[project.scripts]
wireml-runtime = "wireml_runtime.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wireml_runtime"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
